requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=2.0.0
pydantic-settings==2.1.0
python-dotenv>=0.19.0
openai>=1.0.0
slowapi==0.1.8
python-multipart>=0.0.5
lxml[html_clean]>=5.1.0
aiohttp>=3.8.0
websockets>=10.0
Pillow>=9.0.0
numpy>=1.21.0
piexif>=1.1.3
requests>=2.25.0
elevenlabs
newspaper3k>=0.2.8
supabase>=2.0.0
beautifulsoup4>=4.12.0